Remove session client, split up integration and unit test
