+++ src/test/java/io/supertokens/test/DeleteUserAPITest2_8.java
@@ -17,27 +17,15 @@
 package io.supertokens.test;
 
 import com.google.gson.JsonObject;
-import io.supertokens.authRecipe.AuthRecipe;
-import io.supertokens.emailpassword.EmailPassword;
-import io.supertokens.emailpassword.exceptions.ResetPasswordInvalidTokenException;
-import io.supertokens.emailverification.EmailVerification;
-import io.supertokens.emailverification.exception.EmailVerificationInvalidTokenException;
-import io.supertokens.exceptions.UnauthorisedException;
-import io.supertokens.pluginInterface.RECIPE_ID;
+import io.supertokens.Main;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
-import io.supertokens.pluginInterface.emailpassword.UserInfo;
-import io.supertokens.session.Session;
-import io.supertokens.session.info.SessionInformationHolder;
 import io.supertokens.storageLayer.StorageLayer;
 import io.supertokens.test.httpRequest.HttpResponseException;
-import io.supertokens.thirdparty.ThirdParty;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestRule;
-
-import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 
@@ -56,6 +44,36 @@ public class DeleteUserAPITest2_8 {
     }
 
     @Test
+    public void testFailRequestIfBadBody() throws Exception {
+        TestingProcessManager.withProcess(process -> {
+            Main main = process.getProcess();
+
+            if (StorageLayer.getStorage(main).getType() != STORAGE_TYPE.SQL) {
+                return;
+            }
+
+            try {
+                JsonObject bodyWithoutUserId = new JsonObject();
+
+                io.supertokens.test.httpRequest.HttpRequest.sendJsonPOSTRequest(
+                        main,
+                        "",
+                        "http://localhost:3567/user/remove",
+                        bodyWithoutUserId,
+                        1000,
+                        1000,
+                        null,
+                        Utils.getCdiVersion2_8ForTests(),
+                        "");
+
+            } catch (HttpResponseException e) {
+                assertEquals(400, e.statusCode);
+                assertEquals("Http error. Status Code: 400. Message: Field name 'userId' is missing in POST request", e.getMessage());
+            }
+        });
+    }
+
+    @Test
     public void testReturnErrorWhenUserIdDoesntExist() throws Exception {
         TestingProcessManager.withProcess(process -> {
             if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
@@ -66,157 +84,25 @@ public class DeleteUserAPITest2_8 {
             // there's no user in database
 
             // when
-            JsonObject response = deleteUser(process, "inexistent user id");
+            JsonObject requestBody = new JsonObject();
+            requestBody.addProperty("userId", "inexistent user id");
+
+            JsonObject response = io.supertokens.test.httpRequest.HttpRequest.sendJsonPOSTRequest(
+                    process.getProcess(),
+                    "",
+                    "http://localhost:3567/user/remove",
+                    requestBody,
+                    1000,
+                    1000,
+                    null,
+                    Utils.getCdiVersion2_8ForTests(),
+                    ""
+            );
 
             String status = response.get("status").getAsString();
 
             // then
             assertEquals("UNKNOWN_USER_ID_ERROR", status);
         });
-    }
-
-    @Test(expected = UnauthorisedException.class)
-    public void testFailSessionRefreshForDeletedUser() throws Exception {
-        TestingProcessManager.withProcess(process -> {
-            if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
-                return;
-            }
-
-            // given
-            UserInfo user = EmailPassword.signUp(process.getProcess(), "john.doe@example.com", "password");
-
-            SessionInformationHolder sessionWrapper = Session.createNewSession(process.getProcess(), user.id, new JsonObject(), new JsonObject());
-            assert sessionWrapper.refreshToken != null;
-
-            AuthRecipe.deleteUser(process.getProcess(), user.id);
-
-            // when
-            Session.refreshSession(process.getProcess(), sessionWrapper.refreshToken.token, null, false);
-
-            // then should throw
-        });
-    }
-
-    @Test(expected = UnauthorisedException.class)
-    public void testFailSessionVerification() throws Exception {
-        Utils.setValueInConfig("access_token_blacklisting", "true");
-
-        TestingProcessManager.withProcess(process -> {
-            if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
-                return;
-            }
-
-            // given
-            UserInfo user = EmailPassword.signUp(process.getProcess(), "john.doe@example.com", "password");
-
-            SessionInformationHolder sessionWrapper = Session.createNewSession(process.getProcess(), user.id, new JsonObject(), new JsonObject());
-
-            // when
-            AuthRecipe.deleteUser(process.getProcess(), user.id);
-
-            Session.getSession(process.getProcess(), sessionWrapper.session.handle);
-
-            // then should throw
-        });
-    }
-
-    @Test(expected = EmailVerificationInvalidTokenException.class)
-    public void testRemoveEmailValidationTokens() throws Exception {
-        TestingProcessManager.withProcess(process -> {
-            if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
-                return;
-            }
-
-            // given
-            UserInfo user = EmailPassword.signUp(process.getProcess(), "john.doe@example.com", "password");
-
-            String token = EmailVerification.generateEmailVerificationToken(process.getProcess(), user.id, user.email);
-            AuthRecipe.deleteUser(process.getProcess(), user.id);
-
-            // when
-            EmailVerification.verifyEmail(process.getProcess(), token);
-
-            // then should throw
-        });
-    }
-
-    @Test(expected = ResetPasswordInvalidTokenException.class)
-    public void testRemovePasswordResetTokens() throws Exception {
-        TestingProcessManager.withProcess(process -> {
-            if (StorageLayer.getStorage(process.getProcess()).getType() != STORAGE_TYPE.SQL) {
-                return;
-            }
-
-            // given
-            UserInfo user = EmailPassword.signUp(process.getProcess(), "john.doe@example.com", "password");
-
-            String token = EmailPassword.generatePasswordResetToken(process.getProcess(), user.id);
-            AuthRecipe.deleteUser(process.getProcess(), user.id);
-
-            // when
-            EmailPassword.resetPassword(process.getProcess(), token, "newpassword");
-
-            // then should throw
-        });
-    }
-
-    @Test
-    public void testNotReturnRecipeUser() throws Exception {
-        TestingProcessManager.withProcess(process -> {
-            {
-                // given
-                UserInfo user = EmailPassword.signUp(process.getProcess(), "john.doe@example.com", "password");
-
-                long usersCount = AuthRecipe.getUsersCount(process.getProcess(),
-                        new RECIPE_ID[]{RECIPE_ID.EMAIL_PASSWORD});
-
-                assertEquals(1, usersCount);
-
-                // when
-                AuthRecipe.deleteUser(process.getProcess(), user.id);
-
-                // then
-                usersCount = AuthRecipe.getUsersCount(process.getProcess(),
-                        new RECIPE_ID[]{RECIPE_ID.EMAIL_PASSWORD});
-
-                assertEquals(0, usersCount);
-            }
-            {
-                // given
-                ThirdParty.SignInUpResponse response = ThirdParty.signInUp(process.getProcess(), "mockThirdPartyId", "johnDoeId", "john.doe@example.com", true);
-
-                long usersCount = AuthRecipe.getUsersCount(process.getProcess(),
-                        new RECIPE_ID[]{RECIPE_ID.THIRD_PARTY});
-
-                assertEquals(1, usersCount);
-
-                // when
-                AuthRecipe.deleteUser(process.getProcess(), response.user.id);
-
-                // then
-                usersCount = AuthRecipe.getUsersCount(process.getProcess(),
-                        new RECIPE_ID[]{RECIPE_ID.THIRD_PARTY});
-
-                assertEquals(0, usersCount);
-            }
-        });
-    }
-
-    private JsonObject deleteUser(TestingProcessManager.TestingProcess process, String userId) throws IOException,
-            HttpResponseException {
-        JsonObject requestBody = new JsonObject();
-        requestBody.addProperty("userId", userId);
-
-        return io.supertokens.test.httpRequest.HttpRequest.sendJsonPOSTRequest(
-                process.getProcess(),
-                "",
-                "http://localhost:3567/user/remove",
-                requestBody,
-                1000,
-                1000,
-                null,
-                Utils.getCdiVersion2_8ForTests(),
-                ""
-        );
     }
 }
